+++ operatingSystem/FCFSScheduler.h
@@ -11,17 +11,17 @@ class FCFSScheduler : public AScheduler, public IThread
 {
 public:
     // Constructor
-    FCFSScheduler(int nCores, int delayPerExec);
+    FCFSScheduler(int nCores, long long delayPerExec);
 
 	// Instantiates core list based on given number of cores
 	void instantiateCoreList();
 
 	// Runs the actual scheduler
 	void run() override;
-	void runQuantum(int timeQuantum) override;
+	void runQuantum(long long timeQuantum) override;
 	void init() override;
 	void execute() override;
-	void executeQuantum(int timeQuantum) override;
+	void executeQuantum(long long timeQuantum) override;
 
 	void delay(int coreID);
 };
